Fixed unit_price and cantidad validation

+++ Glastore/models/product/__init__.py
@@ -23,7 +23,7 @@ class Product(db.Model, MyModel):
     material = Column(String(100), nullable=False, unique=False, default="")
     cristal = Column(String(100), nullable=False, unique=False, default="")
     acabado = Column(String(100), nullable=False, unique=False, default="")
-    unit_price = Column(Float, nullable=False, default=0)
+    unit_price = Column(Integer, nullable=False, default=0)
     medidas = Column(String(50), nullable=False, unique=False, default="")
     cantidad = Column(Integer, nullable=False, default=0)
     total = Column(Float, nullable=False, default=0)
